requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
idna==2.9
lxml==4.5.0
mysql-connector-python==8.0.19
numpy==1.18.1
pandas==1.0.1
protobuf==3.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
xlrd==1.2.0
